make distro page more obvious
i got to the distro page through the cro documentation and realized that if i
wasn't aware about the debian package, i wouldn't had find it.

this is an attempt to fix it.

https://cro.services/docs/intro/getstarted points to
https://perl6.org/downloads/ and this page is very confusing at first
because you have this "Installing from binaries" that goes to
"https://rakudo.org/files/" with nothing related to the distro
packaging. You have to be carreful to notice in the menu (so not in the
content you're reading), there is this little text entry under a big icon,
that finally give you a pointer to the holy graal which is
https://perl6.org/downloads/others.html.