Initial Raku changes
Now that raku.org no longer redirects, I think it's time to start
changing references to Perl 6. This is the first pass and I manually
corrected just some of the pages. It is a bit hard because there are
many different variations (with space, with no space, with unicode
nbsp, with html entity), and sometimes it is part of “Rakudo Perl 6”,
which is now simply Rakudo.

Not submitting it as a PR because as a PR it won't be finished
ever (some domains are not up yet, some books are not renamed, etc.),
I think just tweaking as much as we can right now is the way to go.