Trim down the /download page
The previous download page gave lots of information on different projects.
Some of that information was stale. Keeping it up to date is difficult.
Thus change it to be simpler and don't try to duplicate information that
runs the risk of turning stale. Also use the chance to educate the reader
a bit about the separation of specification and implementation. That helps
to understand why one has to go to a different website to download the
thing.