Remove mention of the Perl community
It has been almost 2 years since the name change.  People are starting to be confused about the mention of Perl on the home page.  Removing it feels correct in the sense that Raku is becoming more and more a thing on its own.  Also add "Programming Language" and "open source" for better SEO